Move bot code into bot.js
